pyproject.toml
[project]
name = "md2wxhtml"
version = "0.1.14"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "pygments",
    "markdown",
    "premailer",
    "beautifulsoup4",
    "pillow",
]

[project.scripts]
md2wxhtml = "md2wxhtml.main:main"
